moved reader,writer,subscriber to driver-base module
