+++ integrations/apache-camel/src/test/java/org/apache/plc4x/camel/MockDriver.java
@@ -21,7 +21,7 @@ package org.apache.plc4x.camel;
 import org.apache.plc4x.java.spi.PlcDriver;
 import org.apache.plc4x.java.api.authentication.PlcAuthentication;
 import org.apache.plc4x.java.api.PlcConnection;
-import org.apache.plc4x.java.spi.PlcSubscriber;
+import org.apache.plc4x.java.base.messages.PlcSubscriber;
 import org.apache.plc4x.java.api.messages.*;
 import org.apache.plc4x.java.base.messages.DefaultPlcSubscriptionResponse;
 import org.apache.plc4x.java.base.messages.InternalPlcSubscriptionRequest;
