+++ plc4j/protocols/driver-bases/base/src/main/java/org/apache/plc4x/java/base/messages/PlcReader.java
@@ -16,25 +16,25 @@ KIND, either express or implied.  See the License for the
 specific language governing permissions and limitations
 under the License.
 */
-package org.apache.plc4x.java.spi;
+package org.apache.plc4x.java.base.messages;
 
-import org.apache.plc4x.java.api.messages.PlcWriteRequest;
-import org.apache.plc4x.java.api.messages.PlcWriteResponse;
+import org.apache.plc4x.java.api.messages.PlcReadRequest;
+import org.apache.plc4x.java.api.messages.PlcReadResponse;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.function.Consumer;
 
 /**
- * Interface implemented by all PlcConnections that are able to write to remote resources.
+ * Interface implemented by all PlcConnections that are able to read from remote resources.
  */
-public interface PlcWriter {
+public interface PlcReader {
 
     /**
-     * Writes a given value to a PLC.
+     * Reads a requested value from a PLC.
      *
-     * @param writeRequest object describing the type, location and value that whould be written.
-     * @return a {@link CompletableFuture} giving async access to the response of the write operation.
+     * @param readRequest object describing the type and location of the value.
+     * @return a {@link CompletableFuture} giving async access to the returned value.
      */
-    CompletableFuture<PlcWriteResponse> write(PlcWriteRequest writeRequest);
+    CompletableFuture<PlcReadResponse> read(PlcReadRequest readRequest);
 
 }
